OCR/requirements.txt
easyocr==1.7.1
matplotlib==3.7.2
nltk==3.8.1
numpy==1.24.3
opencv_contrib_python==4.6.0.66
opencv_python==4.8.0.74
opencv_python_headless==4.9.0.80
paddleocr==2.8.1
pandas==2.1.3
Pillow==10.2.0
pytesseract==0.3.13
Requests==2.31.0
selenium==4.21.0
torch==2.1.2
torchvision==0.16.2
